+++ ui/src/index.js
@@ -1,12 +1,13 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './css/index.css';
-import App from './app';
+import './app/css/index.css';
+import {PlaygroundApp} from './app';
 import reportWebVitals from './reportWebVitals';
+
 
 ReactDOM.render(
   <React.StrictMode>
-    <App />
+    <PlaygroundApp />
   </React.StrictMode>,
   document.getElementById('root')
 );
